requirements.txt
feedparser>=6.0
httpx>=0.27
trafilatura>=1.12
anthropic>=0.40
python-frontmatter>=1.1
python-slugify>=8.0
pyyaml>=6.0
beautifulsoup4>=4.12
lxml_html_clean>=0.4
